+++ ngmix/observation.py
@@ -674,7 +674,7 @@ class Observation(MetadataMixin):
 
         return Isum, Vsum, Npix
 
-    def copy(self):
+    def copy(self, memo=None):
         """
         make a copy of the observation
         """
@@ -709,7 +709,7 @@ class Observation(MetadataMixin):
         else:
             mfrac = None
 
-        meta = copy.deepcopy(self.meta)
+        meta = copy.deepcopy(self._meta, memo=memo)
 
         return Observation(
             self.image.copy(),
@@ -725,6 +725,55 @@ class Observation(MetadataMixin):
             store_pixels=self._store_pixels,
             ignore_zero_weight=self._ignore_zero_weight,
         )
+
+    def __copy__(self):
+        return self.copy()
+
+    def __deepcopy__(self, memo):
+        result = self.copy(memo=memo)
+        memo[id(self)] = result
+        return result
+
+    def __eq__(self, obs):
+        if not isinstance(obs, Observation):
+            raise ValueError(f'expected Observation, got {type(obs)}')
+
+        # attributes that all share
+        if self.meta != obs.meta:
+            return False
+
+        # image attributes
+        attrs = (
+            ('image', 'array'),
+            ('weight', 'array'),
+            ('bmask', 'array'),
+            ('ormask', 'array'),
+            ('mfrac', 'array'),
+            ('noise', 'array'),
+            ('psf', 'obj'),
+            ('gmix', 'obj'),
+            ('jacobian', 'obj'),
+            ('meta', 'obj'),
+        )
+        for attr, atype in attrs:
+            has = f'has_{attr}'
+            if not hasattr(self, has):
+                # these are probably required
+                self_has = obs_has = True
+            else:
+                self_has = getattr(self, has)()
+                obs_has = getattr(self, has)()
+            if self_has or obs_has:
+                if self_has and obs_has:
+
+                    self_data = getattr(self, attr)
+                    obs_data = getattr(obs, attr)
+                    if not np.all(self_data == obs_data):
+                        return False
+                else:
+                    return False
+
+        return True
 
     @property
     def store_pixels(self):
@@ -889,6 +938,35 @@ class ObsList(list, MetadataMixin):
 
         return Isum, Vsum, Npix
 
+    def copy(self, memo=None):
+        """
+        copy all the data into a new ObsList
+        """
+        new_obslist = ObsList(meta=copy.deepcopy(self._meta, memo))
+        for obs in self:
+            new_obslist.append(obs.copy(memo=memo))
+        return new_obslist
+
+    def __copy__(self):
+        return self.copy()
+
+    def __deepcopy__(self, memo):
+        result = self.copy(memo=memo)
+        memo[id(self)] = result
+        return result
+
+    def __eq__(self, obslist):
+        if not isinstance(obslist, ObsList):
+            raise ValueError(f'expected ObsList, got {type(obslist)}')
+
+        if len(self) != len(obslist):
+            return False
+
+        for self_obs, obs in zip(self, obslist):
+            if self_obs != obs:
+                return False
+        return True
+
     def __setitem__(self, index, obs):
         """
         over-riding this for type safety
@@ -978,6 +1056,37 @@ class MultiBandObsList(list, MetadataMixin):
             Npix += tNpix
 
         return Isum, Vsum, Npix
+
+    def copy(self, memo=None):
+        """
+        copy all the data into a new MultiBandObsList
+        """
+        new_mbobs = MultiBandObsList(meta=copy.deepcopy(self._meta, memo=memo))
+
+        for obslist in self:
+            new_mbobs.append(obslist.copy(memo=memo))
+
+        return new_mbobs
+
+    def __copy__(self):
+        return self.copy()
+
+    def __deepcopy__(self, memo):
+        result = self.copy(memo=memo)
+        memo[id(self)] = result
+        return result
+
+    def __eq__(self, mbobs):
+        if not isinstance(mbobs, MultiBandObsList):
+            raise ValueError(f'expected MultiBandObsList, got {type(mbobs)}')
+
+        if len(self) != len(mbobs):
+            return False
+
+        for self_obslist, obslist in zip(self, mbobs):
+            if self_obslist != obslist:
+                return False
+        return True
 
     def __setitem__(self, index, obs_list):
         """